+++ ninja-core/src/main/java/ninja/jpa/JpaModule.java
@@ -24,6 +24,7 @@ import static com.google.inject.matcher.Matchers.annotatedWith;
 import static com.google.inject.matcher.Matchers.any;
 import com.google.inject.persist.jpa.JpaPersistModule;
 import java.util.Properties;
+import javax.persistence.EntityManager;
 
 public class JpaModule extends AbstractModule {
     
@@ -71,9 +72,10 @@ public class JpaModule extends AbstractModule {
             install(new JpaPersistModule(persistenceUnitName).properties(jpaProperties));
             
             
-            UnitOfWorkInterceptor unitOfWorkInterceptor = new UnitOfWorkInterceptor();
-        
-            requestInjection(unitOfWorkInterceptor);
+            UnitOfWorkInterceptor unitOfWorkInterceptor 
+                    = new UnitOfWorkInterceptor(
+                            getProvider(com.google.inject.persist.UnitOfWork.class),
+                            getProvider(EntityManager.class));
 
             // class-level @UnitOfWork
             bindInterceptor(