+++ ninja-core/src/test/java/ninja/jpa/UnitOfWorkInterceptorTest.java
@@ -16,10 +16,13 @@
 
 package ninja.jpa;
 
+import javax.persistence.EntityManager;
 import org.aopalliance.intercept.MethodInvocation;
 import org.junit.Test;
 import static org.junit.Assert.*;
+import org.junit.Before;
 import org.junit.runner.RunWith;
+import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.runners.MockitoJUnitRunner;
@@ -27,31 +30,48 @@ import org.mockito.runners.MockitoJUnitRunner;
 @RunWith(MockitoJUnitRunner.class)
 public class UnitOfWorkInterceptorTest {
     
-
     @Mock
     MethodInvocation methodInvocation;
     
     @Mock
-    com.google.inject.persist.UnitOfWork UnitOfWork;
+    com.google.inject.persist.UnitOfWork unitOfWork;
     
+    @Mock
+    com.google.inject.Provider<com.google.inject.persist.UnitOfWork> unitOfWorkProvider;
+    
+    @Mock
+    EntityManager entityManager;
+        
+    @Mock
+    com.google.inject.Provider<EntityManager> entityManagerProvider;
+    
+    @Before
+    public void before() {
+    
+        Mockito.when(unitOfWorkProvider.get()).thenReturn(unitOfWork);
+        Mockito.when(entityManagerProvider.get()).thenReturn(entityManager);
+       
+    }
     
     @Test
     public void assertThatSimultaneouslyUsingAnnotationOnManyLevelsWorks() throws Throwable {
         
         // only the most outer annotation should open and close stuff...
 
-        UnitOfWorkInterceptor unitOfWorkInterceptor = new UnitOfWorkInterceptor();
-        unitOfWorkInterceptor.unitOfWork = UnitOfWork;
-        // already started...
-        unitOfWorkInterceptor.didWeStartWork.set(Boolean.TRUE);
+        UnitOfWorkInterceptor unitOfWorkInterceptor 
+                = new UnitOfWorkInterceptor(unitOfWorkProvider, entityManagerProvider);
+        
+        // EntityManager already started (not null) therefore we should NOT
+        // start a unitOfWork
+        Mockito.when(entityManagerProvider.get()).thenReturn(entityManager);
         
         // execute method invocation
         unitOfWorkInterceptor.invoke(methodInvocation);
         
         // no unitOfWork begin
-        Mockito.verify(UnitOfWork, Mockito.never()).begin();
+        Mockito.verify(unitOfWork, Mockito.never()).begin();
         // no unitOfWork ended
-        Mockito.verify(UnitOfWork, Mockito.never()).end();
+        Mockito.verify(unitOfWork, Mockito.never()).end();
         
         // but method has been invoked
         Mockito.verify(methodInvocation).proceed();
@@ -64,16 +84,19 @@ public class UnitOfWorkInterceptorTest {
         
         // only the most outer annotation should open and close stuff...
 
-        UnitOfWorkInterceptor unitOfWorkInterceptor = new UnitOfWorkInterceptor();
-        unitOfWorkInterceptor.unitOfWork = UnitOfWork;
+        UnitOfWorkInterceptor unitOfWorkInterceptor 
+                = new UnitOfWorkInterceptor(unitOfWorkProvider, entityManagerProvider);
+        
+        // not started => unitOfWork should be started
+        Mockito.when(entityManagerProvider.get()).thenReturn(null);
         
         // execute method invocation
         unitOfWorkInterceptor.invoke(methodInvocation);
         
         // no unitOfWork begin
-        Mockito.verify(UnitOfWork).begin();
+        Mockito.verify(unitOfWork).begin();
         // no unitOfWork ended
-        Mockito.verify(UnitOfWork).end();
+        Mockito.verify(unitOfWork).end();
         
         // but method has been invoked
         Mockito.verify(methodInvocation).proceed();