+++ ninja-core/src/site/markdown/developer/changelog.md
@@ -1,6 +1,8 @@
 Version X.X.X
 =============
 
+ * 2014-04-13 BugFix #157. @UnitOfWork did not work when nested with
+              @Transactional or other @UnitOfWork annotations. (ra)
  * 2014-04-07 Now caching virtually infinite amount of template files in memory
               as weak references (before it was a max of 250). (ra)